requirements.txt
asgiref==3.7.2
Django==4.2.4
lxml==4.9.3
python-docx==0.8.11
sqlparse==0.4.4
typing_extensions==4.7.1
